Add event creation code adapted from Portal
As in Portal, added a state handler function which will create an event
every time the things we care about happen in noVNC, we should then be
able to add handlers for these to do useful stuff.

Using jQuery for this at the moment; I'm not sure how good an idea this
is mixing this in with React but seems OK currently and wasn't working
without it (but may be able to get this working with more effort).

Also add a button for connecting now, rather than doing this
automatically, for easier testing.