Copy in useful parts of Aviator
Seems better to approach making the Access Manager similar in code
structure and appearance to Aviator by cutting Aviator down rather than
just extracting certain parts; we know that Aviator works currently
and can just remove unneeded parts rather than trying to find and
extract code and work out what is needed to get it working again in
different circumstances (which I was attempting previously).

This commit replaces the previous generated structure with that of
Aviator, including the build/config files, main components needed for
the Aviator layout, and certain other components which will most
probably also be needed in the Access Manager. Most styling is currently
included also; this can be cut down later as needed.

Files are copied in rather than merged as Aviator is currently closed
source and not publicly available, unlike the Access Manager, so this
avoids exposing the rest of the Aviator codebase through the history.
Unfortunately this does have the downside of losing the history but the
Aviator codebase can be examined for this if needed.