Fully load session data before rendering pages
Now load sessions for authenticated clusters, in addition to the
clusters themselves, before rendering the page, so can read info from
state to decide whether to redirect from session page (in future
commit).

This does mean we need more requests now before the page will display,
with one for the clusters plus one for each authenticated cluster's
sessions; if this is an issue this can be optimized later.