Make copy button copy last copied session text
This commit makes the VNC session copy button copy the last copied text
from the VNC session's clipboard to the user's system clipboard.  This
is done using the Clipboard.js library to handle the details of making
this robust and cross-browser compatible.

This is slightly different to how the copy button worked in Portal, and
how I was originally planning to add this functionality to the access
manager, where the copy button makes the next copied text in the session
be copied to the system clipboard.

Using just JavaScript, this does not currently seem possible, as when
text is selected/copied in the VNC session we receive the JavaScript
click event first, and then the noVNC RFB `onClipboard` function is
called.  However, the browser `document.execCommand('copy')` API (used
by Clipboard.js) requires that it be called within a callback occurring
from some user interaction (to avoid malicious and unexpected additions
to the user's clipboard); and since the click event occurs before we
receive the copied text from noVNC we cannot copy the text to the
clipboard when handling this. I tried getting around this limitation by
having the copying of text to the system clipboard occur when handling a
future user interaction, however this seemed too hacky and flaky to work
effectively.

The equivalent feature in Portal uses Flash to access the clipboard,
however I tried to avoid adding this if possible since Flash is known to
be insecure, and many users do not have it enabled.

In any case, the current UI does not seem less intuitive or deficient to
me than the Portal one, and so seems a good alternative.