Add ping cluster route to determine availability
Note that this is a slightly different implementation from that
originally discussed; originally the pinging was intended to occur as
part of the index route, and each cluster would be pinged to determine
whether to include it in the response to the client.

However, after experimenting with doing this I have found that this
normally takes far too long for the clusters to all either respond or
time out, and this can only take longer in a production environment, so
the app takes much too long to initially load and appears unresponsive.
Therefore an alternative way of doing this has been chosen, where we
will ping clusters after the app has been loaded, and update each of
them independently on the client as we get the results.