No longer reload on return to clusters page
Now that we ping clusters as part of the initial data load, this reload
of clusters on returning to the clusters page was causing an issue
where the cluster boxes would show as loading indefinitely. The simplest
fix is to just remove this reloading; users can always just reload the
page themselves if needed; additionally pub/sub may be added sooner or
later anyway.