Use empty string rather than undefined for default
If an option value is set to undefined or null it will act as if no
option value is set, and so use the option text content as the value.
However, if the value is an empty string this will be used as the value,
so the form submit button will become disabled again if the default
option is reselected.