Pass cluster instead of IP as actions parameter
We now need to access other properties of the cluster an action is for,
when only the action is available to us (specifically when generating
daemon unavailable error messages we want to display the cluster name).
Therefore we with need to pass the whole cluster rather than just the IP
in certain actions so we can read the properties we need.

It makes sense to be consistent with doing this in all actions not just
those we need it for right now, and we may also need further properties
in other actions in future.