Display extra info when qdesktop unavailable
Currently this is just added to the disabled compute node selection
option.

I looked into adding this information to a tooltip, which would be a
nicer way to display this; however tooltips do not work for select
options on many browsers (e.g. Chromium on Debian), and I couldn't find
a library which would enable this without some amount of extra tweaking
and setup (tried `react-bootstrap`, `react-select`,
`react-bootstrap-select`) - can always investigate this further if
needed.