Make clipboard tooltip only show on hover again
Although this is apparently worse for accessibility, with the recent
change to this the 'Copied' message only flashed up briefly (I guess
because the button then became unfocused). Probably there is a better
solution to both these issues but changing back for the moment as I
don't know what that is immediately.