Make app work when devtools unavailable again
An earlier commit (fd73fcfe9fb171d2d5619009feeb5f1e89fdcbb9) replaced
the built-in Redux devtools with the Redux devtools extension;
unfortunately this would break in situations where the devtools were not
available. Fortunately this commit has not been deployed yet, but we
should make sure to include this fix as well in the next release.

Thanks @benarmston for bringing this to my attention.