Explicitly forbid eslint-loader from flight-common
This seems to be necessary to stop `eslint-loader` from attempting to
lint within the `flight-common` directory, even though this is within
the `node_modules` directory which we already forbid it from.

My best guess at the moment is that this is because `flight-common` is
currently linked in, so maybe it is seen as being in the directory it is
linked from rather than within `node_modules` and so is linted.