Make loading react and flight-common work
The `resolve.alias.react` settings ensures that all `require('react')`
expressions resolve to the same copy of react.  That is the one in
alces-access-manager's node_modules directory.

The addition to `resolve.root` ensures that we look in the flight-common dist
directory for required files.  This contains the built flight-common module.