npm-shrinkwrap churn
Includes recent updates to dependencies. In order to successfully
generate the shrinkwrap I needed to (temporarily) change the version of
flight-common used in the package.json to `0.0.0` to match the version
NPM thinks it has installed. Is this the correct way to do this so that
the shrinkwrap is useful in future? Should I specify flight-common
differently in the package.json? Who knows, NPM continues to provide us
with mysteries...