Adapt info/error message customizing to use store
The info/error message generator maps have now been moved to be stored
in the redux store, to prevent an issue where inconsistent instances
would be created when they were imported from AAM and flight-common, and
then the messages would not be set up when we tried to display them.

Storing these in one canonical location, the Redux store, prevents this,
however we are still using the un-Reduxy method of updating the same
object within the state to modify the messages (at some point this may
change to doing this the Redux way).

For now, however, the error messages are set up at initial app load,
while the info messages, which are currently only used for VNC copying,
are set up in the component when needed.