Display screenshot in box for each session
The Clusterware session-screenshot handler sends the captured
screenshots for sessions to the AAM server, which saves them locally.

The AAM client now displays each session's screenshot, if present, or
falls back to an image of some static if no screenshot is available
(e.g. if we have not (yet) received the screenshot or the session
screenshot grabber failed for some reason).