Use Puma instead of WEBrick
I thought using a single-threaded server might be exacerbating an issue
which occurs when a launch session request times out (and so all other
requests would be blocked before the timeout occurred). Using Puma
instead doesn't actually seem to improve this, but still seems a good
idea to use it anyway.