Use fork of react-flipcard to prevent app breaking
In the original react-flipcard (which now appears abandoned), when the
flipcard components would unmount errors would sometimes occur which
completely broke the app. This would particularly happen when we were
waiting for a session to launch and this timed out, we then also have
many load session requests stacked up (both the timeout and the stacking
up of session requests are potential issues I've not yet got to the
bottom of); the AAM app would then break entirely due to these errors.

Using this fork at least fixes this so that when launching a session
times out we just see the error message and then return to a working
app, rather than an unusable broken one.

I have also forked the fork to do this, since this is being installed
from Github rather than the NPM registry this would have broken our
build and be difficult to fix if the original fork ever went away.