Handle session launch timing out better
Rather than display a message about the daemon being unavailable, we now
show a more descriptive message about the launch taking an unusually
long time.

This is done by: adding a new DaemonClient error to distinguish a
communication failure from a timeout, handling this error within the
Rails controller to return the more appropriate gateway timeout (504)
status code, and customizing the error message shown for this action and
status code.