Update screenshot displayed as new ones arrive
We now save new screenshots with a random suffix appended, delete any
old screenshots for a session when we receive a new one, and then
include the current screenshot filename for each session in the session
Json; this means the client can just automatically update the screenshot
displayed as the filename changes.

This also removes the need for using a third party library to determine
when a session screenshot is available, or if we should show a fallback
image, as we can simply tell by whether a filename is included in the
session Json.