Add shrinkpack npm script and explicit dependency
Dependency already existed within `node_modules`/`node_shrinkwrap`, from
some other dependency I think, but make this explicit and add npm script
to run it. This is done so not relying on developers having this
installed globally or on it being implicitly included in our
dependencies because of some other dependency.