Add Puma config adapted from ASM
This is needed in production so we can configure the server and set
environment variables to be used in the ASM starter script.

Only real difference from the ASM config
(https://github.com/alces-software/alces-storage-manager/blob/master/config/puma.rb)
is removing establishing an ActiveRecord connection, as AAM has no
database and I think trying to do this was also causing the server to
fail.