Display correct SSH address for clusters
In the AAM config file clusters are specified along with an IP, however
this IP is intended to be one which the AAM server can reach, with no
guarantee that this will be reachable to users viewing the AAM client.

Therefore, rather than using this IP directly, the AAM daemon now
includes the IP it has identified as the cluster's public IP along with
the other data returned to the AAM server, and this is used in the SSH
command instead as it is more likely to work.