Only show SSH access info if know cluster login IP
If we don't have the cluster login IP for some reason, we don't know
what the IP for SSH access should be, so should show nothing.

As part of this change the text referring to SSH access has also been
moved into the SSH access component, so this can also be shown/hidden
with the same logic.