In production use proxy on login node
Rather than on AAM appliance as was done before.

The idea is that we only need a single IP (the cluster login node IP),
with minimal ports open, routable from wherever the AAM is running,
whether that is within the cluster network or externally. All websocket
connections to any node within the cluster are then securely proxied via
this node.