Correct method for determining login node sessions
As far as I can tell, Clusterware sessions' `host` and `access_host`
always seem to be equal; therefore the previous method didn't work.

Instead, we can determine that a session is on the login node if and
only if its host IP is the same as the login node's IP (at least, I hope
so!).